Modified task 9 (and task 8) so that large number of samples can be generated and a dynamically determined amount of them is written to the output file. (This is optimized for logarithmic plotting.)

+++ BaSheet01/Task9/convergence_plot_data.h
@@ -12,6 +12,6 @@
 
 #include "INormalDist.h"
 
-void generate_convergence_data_with(double mu, double sigma, unsigned int size, INormalDistribution & dice);
+void generate_convergence_data_with(double mu, double sigma, long long unsigned int size, INormalDistribution & dice);
 
 #endif /* CONVERGENCE_PLOT_DATA_H_ */
